requirements.txt
numpy
torch
torchvision
datasets
pillow
PyYAML
tqdm
matplotlib
scikit-learn
scipy
pandas
seaborn
tensorflow